+++ src/app/socketTest/page.js
@@ -7,7 +7,7 @@ const URL = "http://localhost:3001";
 
 const socket = io("http://localhost:3001");
 
-export default function Index() {
+export default function SocketTest() {
     const [latestMsg, setLatestMsg] = useState("null");
     const [fieldText, setFieldText] = useState("");
